Rename: VO 디렉터리 생성 후 이동 (#61)
