+++ src/main/java/dev/yogizogi/domain/meokprofile/model/vo/MeokProfileVO.java
@@ -1,7 +1,10 @@
-package dev.yogizogi.domain.meokprofile.model.entity;
+package dev.yogizogi.domain.meokprofile.model.vo;
 
+import dev.yogizogi.domain.meokprofile.model.entity.Intensity;
+import dev.yogizogi.domain.meokprofile.model.entity.Preference;
 import dev.yogizogi.domain.user.model.entity.User;
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.util.Objects;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -22,6 +25,22 @@ public class MeokProfileVO {
     public MeokProfileVO(User user, Intensity intensity, Preference preference) {
         this.intensity = intensity;
         this.preference = preference;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof MeokProfileVO that)) {
+            return false;
+        }
+        return intensity.equals(that.intensity) && preference.equals(that.preference);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(intensity, preference);
     }
 
 }
