+++ src/main/java/dev/yogizogi/domain/review/service/MenuReviewService.java
@@ -7,7 +7,7 @@ import static dev.yogizogi.global.common.code.ErrorCode.NOT_EXIST_REVIEW;
 import dev.yogizogi.domain.menu.exception.NotExistMenuException;
 import dev.yogizogi.domain.menu.model.entity.Menu;
 import dev.yogizogi.domain.menu.repository.MenuRepository;
-import dev.yogizogi.domain.meokprofile.model.entity.MeokProfileVO;
+import dev.yogizogi.domain.meokprofile.model.vo.MeokProfileVO;
 import dev.yogizogi.domain.review.execption.NoPermissionRestaurantException;
 import dev.yogizogi.domain.review.execption.NotExistReviewException;
 import dev.yogizogi.domain.review.model.dto.request.CreateMenuReviewInDto;
@@ -16,7 +16,7 @@ import dev.yogizogi.domain.review.model.dto.response.GetMenuReviewOutDto;
 import dev.yogizogi.domain.review.model.dto.response.GetMenuReviewsOutDto;
 import dev.yogizogi.domain.review.model.entity.MenuReview;
 import dev.yogizogi.domain.review.model.entity.MenuReviewImage;
-import dev.yogizogi.domain.review.model.entity.MenuReviewVO;
+import dev.yogizogi.domain.review.model.vo.MenuReviewVO;
 import dev.yogizogi.domain.review.model.entity.Review;
 import dev.yogizogi.domain.review.repository.MenuReviewRepository;
 import dev.yogizogi.domain.review.repository.ReviewImageRepository;
